+++ compose.yml
@@ -62,7 +62,6 @@ services:
       - node_modules_ghost_post-revisions:/home/ghost/ghost/post-revisions/node_modules:delegated
       - node_modules_ghost_posts-service:/home/ghost/ghost/posts-service/node_modules:delegated
       - node_modules_ghost_prometheus-metrics:/home/ghost/ghost/prometheus-metrics/node_modules:delegated
-      - node_modules_ghost_recommendations:/home/ghost/ghost/recommendations/node_modules:delegated
       - node_modules_ghost_security:/home/ghost/ghost/security/node_modules:delegated
       - node_modules_ghost_tiers:/home/ghost/ghost/tiers/node_modules:delegated
       - node_modules_ghost_webmentions:/home/ghost/ghost/webmentions/node_modules:delegated
@@ -191,7 +190,6 @@ volumes:
   node_modules_ghost_post-revisions: {}
   node_modules_ghost_posts-service: {}
   node_modules_ghost_prometheus-metrics: {}
-  node_modules_ghost_recommendations: {}
   node_modules_ghost_security: {}
   node_modules_ghost_tiers: {}
   node_modules_ghost_webmentions: {}
