+++ ghost/core/core/server/services/recommendations/service/IncomingRecommendationEmailRenderer.ts
@@ -34,4 +34,4 @@ export class IncomingRecommendationEmailRenderer {
             recipient
         });
     }
-};
+}
