+++ ghost/core/core/server/services/recommendations/service/RecommendationMetadataService.ts
@@ -1,3 +1,5 @@
+/* eslint-disable ghost/filenames/match-exported-class */
+
 type OembedMetadata<Type extends string> = {
     version: '1.0',
     type: Type,
