--- ghost/recommendations/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@tryghost/recommendations",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/ghost/recommendations",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
-  "scripts": {
-    "dev": "tsc --watch --preserveWatchOutput --sourceMap",
-    "build": "yarn build:ts",
-    "build:ts": "tsc",
-    "prepare": "tsc",
-    "test:unit": "NODE_ENV=testing c8 --src src --all --check-coverage --100 --reporter text --reporter cobertura -- mocha --reporter dot -r ts-node/register './test/**/*.test.ts'",
-    "test": "yarn test:types && yarn test:unit",
-    "test:types": "tsc --noEmit",
-    "lint:code": "eslint src/ --ext .ts --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .ts --cache"
-  },
-  "files": [
-    "build"
-  ],
-  "devDependencies": {
-    "@tryghost/nql": "0.12.7",
-    "@types/node": "20.14.8",
-    "c8": "8.0.1",
-    "mocha": "10.8.2",
-    "sinon": "15.2.0",
-    "ts-node": "10.9.2",
-    "typescript": "5.8.2"
-  },
-  "dependencies": {
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/in-memory-repository": "0.0.0",
-    "@tryghost/logging": "2.4.18",
-    "@tryghost/mongo-utils": "0.6.2",
-    "@tryghost/tpl": "0.1.32",
-    "knex": "2.4.2"
-  }
-}
